Rename files and update license and docs (#183)
* rename  files

* update comment

* update license

* address comments
